Finished Comment Entity

+++ BlogPostAPI/Migrations/DataContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace BlogPostAPI.Migrations
 
             modelBuilder.Entity("BlogPostAPI.Models.Comment", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<long>("PostId")
                         .HasColumnType("bigint");
